pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blockdownload"

authors = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" }
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" }
]
readme = "README.md"
license = "Apache-2.0"

dependencies = [
  # https://github.com/WolfgangFahl/pyLoDStorage
  "pyLodStorage>=0.14.3",
  # https://pypi.org/project/tqdm/
  "tqdm>=4.66.5"
]
requires-python = ">=3.10"

classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13"
]

dynamic = ["version", "description"]

[project.urls]
Home = "https://github.com/WolfgangFahl/blockdownload"
Documentation = "http://wiki.bitplan.com/index.php/blockdownload"
Source = "https://github.com/WolfgangFahl/blockdownload"

[tool.hatch.version]
path = "bdown/__init__.py"

[project.optional-dependencies]
test = []

[tool.hatch.build.targets.wheel]
only-include = ["bdown"]

[tool.hatch.build.targets.wheel.sources]
"osprojects" = "bdown"


[project.scripts]
dcheck = "bdown.check:main"
blockdownload="bdown.download_cmd:main"
